requirements.txt
streamlit
openai>=1.0.0
python-dotenv
pdfplumber
pydantic
numpy
markdown
qdrant-client